don't repeat yourself and syntax cleanup
First: nice job, this code works!

Look at lines 6 and 29 of this pull request. They are not doing the exact same thing, but they are doing something similar. Is it necessary to have `div` and `mod` appear in two places in this code? Or can they appear only once?

Cleanup: unnecessary parantheses. See the other pull request.